requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
mf2py==1.1.2
pycrypto==2.6.1
PyGObject==3.30.4
python-dotenv==0.14.0
pyxdg==0.25
requests==2.24.0
SecretStorage==2.3.1
six==1.12.0
soupsieve==2.0.1
urllib3==1.25.10
uWSGI==2.0.19.1
webencodings==0.5.1
Werkzeug==1.0.1
